requirements.txt
Flask
requests
Cmake
supabase
openai
langchain
langchain_openai
jq
pydantic
gunicorn
fashion_clip
scipy
python-dotenv